+++ asusctl/src/cli_opts.rs
@@ -1,5 +1,5 @@
 use gumdrop::Options;
-use rog_platform::platform::PlatformPolicy;
+use rog_platform::platform::ThrottlePolicy;
 
 use crate::anime_cli::AnimeCommand;
 use crate::aura_cli::{LedBrightness, LedPowerCommand1, LedPowerCommand2, SetAuraBuiltin};
@@ -60,7 +60,7 @@ pub struct ProfileCommand {
     pub profile_get: bool,
 
     #[options(meta = "", help = "set the active profile")]
-    pub profile_set: Option<PlatformPolicy>,
+    pub profile_set: Option<ThrottlePolicy>,
 }
 
 #[derive(Options)]